Move the test fixtures to global utils
Move the test fixtures to global utils so that they can be
shared between different applications.
